dbeaver/dbeaver-ee#1981 Refactor blocking remote call execution with progress
